feat(dashboard): update sign in and sign up pages
