+++ dashboard/src/pages/signin/email.tsx
@@ -3,13 +3,12 @@ import NavLink from '@/components/common/NavLink';
 import UnauthenticatedLayout from '@/components/layout/UnauthenticatedLayout';
 import Box from '@/ui/v2/Box';
 import Button from '@/ui/v2/Button';
-import Divider from '@/ui/v2/Divider';
 import Input, { inputClasses } from '@/ui/v2/Input';
 import Text from '@/ui/v2/Text';
 import { getToastStyleProps } from '@/utils/settings/settingsConstants';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { styled } from '@mui/material';
-import { useSignUpEmailPassword } from '@nhost/nextjs';
+import { useSignInEmailPassword } from '@nhost/nextjs';
 import type { ReactElement } from 'react';
 import { useEffect } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
@@ -19,7 +18,6 @@ import * as Yup from 'yup';
 const validationSchema = Yup.object({
   email: Yup.string().label('Email').email().required(),
   password: Yup.string().label('Password').required(),
-  displayName: Yup.string().label('Name').required(),
 });
 
 export type EmailSignUpFormValues = Yup.InferType<typeof validationSchema>;
@@ -32,14 +30,13 @@ const StyledInput = styled(Input)({
 });
 
 export default function EmailSignUpPage() {
-  const { signUpEmailPassword, error } = useSignUpEmailPassword();
+  const { signInEmailPassword, error } = useSignInEmailPassword();
 
   const form = useForm<EmailSignUpFormValues>({
     reValidateMode: 'onSubmit',
     defaultValues: {
       email: '',
       password: '',
-      displayName: '',
     },
     resolver: yupResolver(validationSchema),
   });
@@ -52,23 +49,17 @@ export default function EmailSignUpPage() {
     }
 
     toast.error(
-      error?.message || 'An error occurred while signing up. Please try again.',
+      error?.message || 'An error occurred while signing in. Please try again.',
       getToastStyleProps(),
     );
   }, [error]);
 
-  async function handleSubmit({
-    email,
-    password,
-    displayName,
-  }: EmailSignUpFormValues) {
+  async function handleSubmit({ email, password }: EmailSignUpFormValues) {
     try {
-      await signUpEmailPassword(email, password, {
-        displayName,
-      });
+      await signInEmailPassword(email, password);
     } catch {
       toast.error(
-        'An error occurred while signing up. Please try again.',
+        'An error occurred while signing in. Please try again.',
         getToastStyleProps(),
       );
     }
@@ -81,7 +72,7 @@ export default function EmailSignUpPage() {
         component="h1"
         className="text-center text-3.5xl font-semibold lg:text-4.5xl"
       >
-        Sign Up with Email
+        Sign In
       </Text>
 
       <Box className="grid grid-flow-row gap-4 rounded-md border bg-transparent p-6 lg:p-12">
@@ -90,22 +81,6 @@ export default function EmailSignUpPage() {
             onSubmit={handleSubmit}
             className="grid grid-flow-row gap-4 bg-transparent"
           >
-            <StyledInput
-              {...register('displayName')}
-              id="displayName"
-              placeholder="Name"
-              inputProps={{ min: 2, max: 128 }}
-              spellCheck="false"
-              autoCapitalize="none"
-              type="text"
-              label="Name"
-              hideEmptyHelperText
-              fullWidth
-              autoComplete="off"
-              error={!!formState.errors.displayName}
-              helperText={formState.errors.displayName?.message}
-            />
-
             <StyledInput
               {...register('email')}
               id="email"
@@ -136,6 +111,14 @@ export default function EmailSignUpPage() {
               helperText={formState.errors.password?.message}
             />
 
+            <NavLink
+              href="/reset-password"
+              color="white"
+              className="justify-self-start font-semibold"
+            >
+              Forgot password?
+            </NavLink>
+
             <Button
               className="!bg-white !text-black disabled:!text-black disabled:!text-opacity-60"
               size="large"
@@ -143,48 +126,23 @@ export default function EmailSignUpPage() {
               loading={formState.isSubmitting}
               type="submit"
             >
-              Sign Up with Email
+              Sign In
             </Button>
 
             <Text color="secondary" className="text-center">
               or{' '}
-              <NavLink color="white" className="font-semibold" href="/signup">
-                sign up with GitHub
+              <NavLink color="white" className="font-semibold" href="/signin">
+                sign in with GitHub
               </NavLink>
             </Text>
           </Form>
         </FormProvider>
-
-        <Divider />
-
-        <Text color="secondary" className="text-center text-sm">
-          By clicking continue, you agree to our{' '}
-          <NavLink
-            href="https://nhost.io/legal/terms-of-service"
-            target="_blank"
-            rel="noopener noreferrer"
-            className="font-semibold"
-            color="white"
-          >
-            Terms of Service
-          </NavLink>{' '}
-          and{' '}
-          <NavLink
-            href="https://nhost.io/legal/privacy-policy"
-            target="_blank"
-            rel="noopener noreferrer"
-            className="font-semibold"
-            color="white"
-          >
-            Privacy Policy
-          </NavLink>
-        </Text>
       </Box>
 
       <Text color="secondary" className="text-center text-base lg:text-lg">
-        Already have an account?{' '}
-        <NavLink href="/signin" color="white">
-          Sign in
+        Don&apos;t have an account?{' '}
+        <NavLink href="/signup" color="white">
+          Sign Up
         </NavLink>
       </Text>
     </>
@@ -192,5 +150,5 @@ export default function EmailSignUpPage() {
 }
 
 EmailSignUpPage.getLayout = function getLayout(page: ReactElement) {
-  return <UnauthenticatedLayout title="Sign Up">{page}</UnauthenticatedLayout>;
+  return <UnauthenticatedLayout title="Sign In">{page}</UnauthenticatedLayout>;
 };
